Refactor: Get query data in KeywordRecommends
